+++ TiendaE-comerce/src/main/java/com/mintic/TiendaEcomerce/dao/ClientesDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mintic.TiendaEcomerce.model.Clientes;
 
-public interface ClientesDao extends JpaRepository<Clientes, Integer>
+public interface ClientesDao extends JpaRepository<Clientes, Long>
 {
 
 }
